+++ src/constants/soundNotification.ts
@@ -1,3 +1,5 @@
+import { Sounds } from 'Libs/Sound'
+
 /**
  * Sound notifications.
  */
@@ -5,6 +7,15 @@ enum SoundNotification {
   Message,
   Mention,
   Whisper,
+}
+
+/**
+ * Sound notification to audio sound mapping.
+ */
+export const SoundNotificationAudioMap = {
+  [SoundNotification.Mention]: Sounds.Notification,
+  [SoundNotification.Message]: Sounds.Message,
+  [SoundNotification.Whisper]: Sounds.Notification,
 }
 
 export default SoundNotification