+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ### 🚀 New Feature
 
 - Available logs can now be exported to a file or your clipboard from the Tools menu. Pick what you want to include and click that button!
-- Add a new notification (disabled by default) to play a sound for each message.
+- Notifications:
+  - Add a new notification (disabled by default) to play a sound for each message.
+  - The volume of each sound can now be controlled independently.
 
 ### 🐛 Bug Fix